Added progress reporters and added features
Progress report types to double.
Added a faster math method for rounding (2ns faster but still)
Added a boolean for telling whether a package is encrypted.
Fixed some minor bugs.

+++ src/Application/Common/Interfaces/IAesEncryptionService.cs
@@ -13,7 +13,7 @@
         /// <param name="progressReportInterval">The interval in milliseconds for updating the progress instances when present.</param>
         /// <param name="cancellationToken">A cancellation token to cancel the decryption process.</param>
         /// <returns><see langword="true"/> when successful, <see langword="false"/> otherwise.</returns>
-        bool DecryptCbc(byte[] input, int dataSize, out byte[] output, uint[] key, IProgress<int>? progress = null,
+        bool DecryptCbc(byte[] input, int dataSize, out byte[] output, uint[] key, IProgress<double>? progress = null,
             int progressReportInterval = 100, CancellationToken cancellationToken = default);
 
         /// <summary>
@@ -26,7 +26,7 @@
         /// <param name="progressReportInterval">The interval in milliseconds for updating the progress instances when present.</param>
         /// <param name="cancellationToken">A cancellation token to cancel the encryption process.</param>
         /// <returns><see langword="true"/> when successful, <see langword="false"/> otherwise.</returns>
-        bool EncryptCbc(byte[] input, out byte[] output, uint[] key, IProgress<int>? progress = null,
+        bool EncryptCbc(byte[] input, out byte[] output, uint[] key, IProgress<double>? progress = null,
             int progressReportInterval = 100, CancellationToken cancellationToken = default);
 
         /// <summary>
