+++ src/Application/PathEntries/Queries/ExportPathEntriesQuery.cs
@@ -29,7 +29,7 @@ namespace ResourcePackerGUI.Application.PathEntries.Queries
         /// <summary>
         /// An optional progress instance to keep track of the writing process.
         /// </summary>
-        public IProgress<int>? Progress { get; init; }
+        public IProgress<double>? Progress { get; init; }
 
         /// <summary>
         /// The interval in milliseconds for updating the progress instances when present.
