+++ src/Application/Resources/Queries/ExportResourcesQuery.cs
@@ -37,7 +37,7 @@ namespace ResourcePackerGUI.Application.Resources.Queries
         /// <summary>
         /// An optional progress instance to keep track of the export process.
         /// </summary>
-        public IProgress<int>? Progress { get; init; }
+        public IProgress<double>? Progress { get; init; }
 
         /// <summary>
         /// The interval in milliseconds for updating the progress instances when present.
